pyproject.toml
[tool.poetry]
name = "drf-fat-models"
version = "0.1.0"
description = "Place business logic in one place at the object-level, use automatically from anywhere: API, Django admin and forms."
license = "MIT"
authors = ["ZeBrains <a.chernousov@zebrains.team>"]
maintainers = ["Andrey Chernousov <a.chernousov@zebrains.team>"]
readme = "README.md"
homepage = "https://github.com/zebrainsteam/drf-fat-models"
repository = "https://github.com/zebrainsteam/drf-fat-models"
keywords=["django", "drf", "fat model", "model", "business logic"]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 2.2",
    "Framework :: Django :: 3.0",
    "Framework :: Django :: 3.1",
    "Framework :: Django :: 3.2",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.6"
django = ">=2.2"
djangorestframework = ">=3.11"

[tool.poetry.dev-dependencies]
flake8 = "*"
flake8-isort = "*"
flake8-docstrings = "*"
flake8-builtins = "*"
flake8-quotes = "*"
flake8-bugbear = "*"
flake8-coding = "*"
flake8-comprehensions = "*"
flake8-eradicate = "*"
flake8-pep3101 = "*"
flake8-string-format = "*"
flake8-print = "*"
flake8-super-call = "*"
flake8-django = "*"
bandit = "*"
autopep8 = "*"
black = "*"
isort = "*"
docformatter = "*"
pre-commit = "*"
pytest = "*"
pytest-xdist = "*"
pytest-cov = "*"
pytest-django = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.autopep8]
max_line_length = 88
